requirements.txt
django-viewcomponent == 1.0.5
